Add support for processing nonlinearly binned files. Various routines in binning.py were written to throw an error on nonlinear files, but as far as I can tell there is no reason for this to happen because they function perfectly well, since they only operate on transmitted bins anyway.
